feat: create a game context
