+++ frontend/src/app/(private)/game/play/template.tsx
@@ -1,0 +1,9 @@
+import { GameProvider } from "@/contexts/GameContext";
+
+export default function Template({ children }: { children: React.ReactNode }) {
+    return (
+        <GameProvider>
+            <div>{children}</div>
+        </GameProvider>
+    )
+  }
